requirements.txt
python-dotenv==0.20.*
pytest==7.1.*
requests==2.28.*
sentry-sdk[fastapi]==1.10.*
alembic==1.8.*
gevent==22.10.*
feedparser==6.0.10
sqlalchemy==2.0.*
psycopg>=3.1.*
python-dateutil==2.8.*
mediacloud-metadata>=0.7.6
fastapi==0.79.*
Jinja2==3.1.*
uvicorn==0.18.*
statsd_client==1.0.*
setproctitle==1.3.*
PyYAML==6.0.*
mediacloud>=4.0.1
